Define a tibble period varset

A function to join a list of tibbles, by a single population_member_id, using either the inner_join or left_join functions from dplyr.

define_varset(var_defs = list(), population_member_id = "id_referral_visit",
  multi_join_type = "inner")

Arguments

var_defs
A list of tibble variables (events, periods, or attributes).

population_member_id
A unique identifier for each 'member' of the population of interest. Must be present in each tibble.

multi_join_type
The name of the join you wish to make. Currently limited to 'inner' or 'left'. Defaults is 'inner'.