Create a vector of 24 hour Intervals lubridate from oliver_replica write a single file

list_holidays_and_weekends(from_date = "01-01-2011", to_date = "01-01-2111",
  country_filter = "usa", region_filter = "Washington")

Arguments

from_date
A character string specifying the start date of the Interval vector in the format " to_dateA character string specifying the stop date of the Interval vector in the format " country_filterA character string specifying country of the holidays as required by the Enrico service. Defaults to 'usa'. region_filterA character string specifying region, of the country, of the holidays as required by the Enrico service. Defaults to 'Washington'. A vector of Interval objects. This function is used to generate a vector of holiday and weekend Interval objects as defined by lubridate. It requires an active internet connection as it makes use of the Enrico service available at http://www.kayaposoft.com/enrico/ to determine holidays for a given locality. This is used to exclude weekends and holidays from Duration and Period calculations. list_holidays_and_weekends() Joe Mienko, mienkoja@uw.edu